Fix Typescript export definitions
Typescript does not support .mjs file, so the module that ends up being
imported is the commonJS version (`dist/unfetch.js`). Since it does not contain a
"default" export entry, using it in Typescript results in a runtime
error: `TypeError: isomorphic_unfetch_1.default is not a function`

Fix by reverting to `export =`. To use in Typescript, use the form
`import * as fetch from "isomorphic-unfetch"`.